Driver commissions with invalid customers
Update the driver commission history model to just include the customer
so it is a left join which will handle funny (non customer id) data
in the CustomerId of the Trucking_Drivers_Commissions_History schema.

This could be specific to our (Western Milling) Grossman/Relativity
schemas but for our use this isn't a problem for now.

Closes #61