Add item price to order response [TP#558]
Add the item price to the order response so that clients have access
to item pricing information from Grossman.

Also took this change to sneak adding in the CarrierId and the ShipTime
to the order response as these will also be needed.

Updated some configuration so that the conditional logger is decided
in the staging and production environments rather than in the application
so that the development and tests environments aren't affected.

You'll note that a couple of existing migrations were changed, with a
couple of fields made not nullable... normally I wouldn't do this but
since those migrations are used by the test suite only I deemed it
to be of no harm.

Also renamed the order request test as it was wrongly named and thus
not being included, which resulted in some work to get those tests
passing.