Fix incorrect order status returned [TP#607]
It would seem that Grossman does not update the order status field
in the InvCustomerOrders_CrossReference schema so this becomes out
of date and thus the Order::Reference status is incorrect.

To handle this the response has been updated to use the status
from the Order model (InvCustomerOrders_Open_MainInfo schema) which
is the up-to data order status.