Add missing fields to contract response [TP#887]
The contract response is missing the location which is pretty crucial
for contract uniqueness, and there is no inv contract id which is
a convenience composite of contract number and sub.

Also, the contract type is using the old single character and not the
more verbose value.

WIP, update the request tests to reflect the new response we expect.