Fix contract parameter type mismatch [TP#895]
Update the contract scope SQL to wrap inventory contract id values
in single quotes as this is a string field in Grossman/Relativity.

This would not fail when testing with local (non Relativity) datasources
since they perform implicit type conversion so `SELECT 1 == '1'` is true,
however this is not the case for Relativity. For now I propose additional
pre merge tests against a Relativity datasource.