Fix issues with querying production data
Update some relationships to provide commodity and contract details
and filters through other means than the contract as I was unable to
find contracts that matched the ContractId of pick up orders.

Downgrade Rails back due to errors:

```
NoMethodError: undefined method `type' for 1:Fixnum
```

No real benefits to using the latest Rails at this time so it was
more practical to revert rather than try and resolve the issue.