Tweak pick up contract query methods
Some strange behavior is exhibited when querying pick up orders and
requiring the contract details returned in a single query.

Long story short, some queries (not those using "scope" methods)
would not return the contract in a single query and thus blow up when
attempting to fetch the contracts with a LIMIT. A joins to the contract
association was needed to ensure the contract details were returned
in a single query.

Due to this is consistent behavior it is expected that the client will
be responsible for indicating that they require contract details by
using the `contract_present` ransack scope.

Ultimately I recommend that contract details are not returned by the
pick up order response and the contract details are fetched with another
request to return contracts by a list of  ContractId or Inv_ContractId
values.

The pick up contract association should just be used for queries that
involve contract related information, i.e. balance, presence.