Rescue connection timeout in health check ID:2349
https://westernmilling.tpondemand.com/entity/2349

We seem to get these errors consistently when Relativity is down (or so
slow as to appear down), so this change rescues them and reports it as a
failed health check.

Also adds a timeout as spec'd in the story, but I could not find a way
to test this even with Timecop. I did manually test that the timeout
works.